Fix x64 builds

+++ core/flavors/flavor_wcs_comlib.cpp
@@ -86,13 +86,13 @@ namespace la::flavors
 
 			//read the tag
 			{
-				line.sectionTag.offset = walker - line.data.start;
+				line.sectionTag.offset = static_cast<uint16_t>(walker - line.data.start);
 				line.sectionTag.size = 0;
 
 				while ((walker < walkerEnd) && (walker[0] != ':'))
 					walker++;
 
-				line.sectionTag.size = walker - line.data.start - line.sectionTag.offset;
+				line.sectionTag.size = static_cast<uint32_t>(walker - line.data.start - line.sectionTag.offset);
 
 				walker++; //ignore ':'
 			}
@@ -103,13 +103,13 @@ namespace la::flavors
 					return false;
 
 				walker++; //skip space
-				line.sectionMethod.offset = walker - line.data.start;
+				line.sectionMethod.offset = static_cast<uint16_t>(walker - line.data.start);
 				line.sectionMethod.size = 0;
 
 				while ((walker < walkerEnd) && (*walker != '|'))
 					walker++;
 
-				line.sectionMethod.size = walker - line.data.start - line.sectionMethod.offset;
+				line.sectionMethod.size = static_cast<uint32_t>(walker - line.data.start - line.sectionMethod.offset);
 
 				if ((line.sectionMethod.size < 1) || (walker[-1] != ' ')) //last caracter before '|' must be a space
 					return false;
@@ -124,13 +124,13 @@ namespace la::flavors
 					return false;
 
 				walker++; //skip space
-				line.sectionMsg.offset = walker - line.data.start;
+				line.sectionMsg.offset = static_cast<uint16_t>(walker - line.data.start);
 				line.sectionMsg.size = 0;
 
 				while ((walker < walkerEnd) && (walker[0] != '|'))
 					walker++;
 
-				line.sectionMsg.size = walker - line.data.start - line.sectionMsg.offset;
+				line.sectionMsg.size = static_cast<uint32_t>(walker - line.data.start - line.sectionMsg.offset);
 
 				if (walker < walkerEnd) //the message *can* be the last thing in the line
 				{
@@ -149,8 +149,8 @@ namespace la::flavors
 					return false;
 
 				walker++;
-				line.sectionParams.offset = walker - line.data.start;
-				line.sectionParams.size = walkerEnd - walker; //end of the line
+				line.sectionParams.offset = static_cast<uint16_t>(walker - line.data.start);
+				line.sectionParams.size = static_cast<uint32_t>(walkerEnd - walker); //end of the line
 			}
 
 			//:)
